Make AutoArrangeChildren into a Scene extension method
